+++ mainapp/graphql/private/resolvers/permissions.resolvers.go
@@ -8,266 +8,12 @@ import (
 	"dataplane/auth_permissions"
 	"dataplane/database"
 	"dataplane/database/models"
-	privategraphql "dataplane/graphql/private"
 	"dataplane/logging"
 	"errors"
 	"os"
 
-	"github.com/google/uuid"
 	"gorm.io/gorm"
 )
-
-func (r *mutationResolver) CreateAccessGroup(ctx context.Context, environmentID string, name string, description *string) (string, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return "", errors.New("Requires permissions.")
-	}
-
-	e := models.PermissionsAccessGroups{
-		AccessGroupID: uuid.New().String(),
-		Name:          name,
-		Description:   *description,
-		EnvironmentID: environmentID,
-		Active:        true,
-	}
-
-	err := database.DBConn.Create(&e).Error
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return "", errors.New("Add access group database error.")
-	}
-
-	return e.AccessGroupID, nil
-}
-
-func (r *mutationResolver) UpdateAccessGroup(ctx context.Context, input *privategraphql.AccessGroupsInput) (string, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-	environmentID := input.EnvironmentID
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return "", errors.New("Requires permissions.")
-	}
-
-	err := database.DBConn.Where("access_group_id = ?", input.AccessGroupID).Updates(models.PermissionsAccessGroups{
-		Name:          input.Name,
-		Description:   input.Description,
-		Active:        input.Active,
-		EnvironmentID: input.EnvironmentID,
-	}).Error
-
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return "", errors.New("update user database error")
-	}
-
-	response := "success"
-	return response, nil
-}
-
-func (r *mutationResolver) ActivateAccessGroup(ctx context.Context, accessGroupID string, environmentID string) (string, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return "", errors.New("Requires permissions.")
-	}
-
-	// Check if access group alredy active
-	p := models.PermissionsAccessGroups{}
-
-	err := database.DBConn.Where("access_group_id = ?", accessGroupID).First(&p).Error
-
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return "", errors.New("Retrive me database error.")
-	}
-
-	if p.Active == true {
-		return "", errors.New("User is already active.")
-	}
-
-	// Activate user
-	err = database.DBConn.Where(&models.PermissionsAccessGroups{AccessGroupID: accessGroupID}).Select("active", false).
-		Updates(models.PermissionsAccessGroups{Active: true}).Error
-
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return "", errors.New("Activate access group database error.")
-	}
-
-	response := "Access group activated"
-	return response, nil
-}
-
-func (r *mutationResolver) DeactivateAccessGroup(ctx context.Context, accessGroupID string, environmentID string) (string, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return "", errors.New("Requires permissions.")
-	}
-
-	// Check if access group alredy inactive
-	p := models.PermissionsAccessGroups{}
-
-	err := database.DBConn.Where("access_group_id = ?", accessGroupID).First(&p).Error
-
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return "", errors.New("Retrive me database error.")
-	}
-
-	if p.Active == false {
-		return "", errors.New("User is already inactive.")
-	}
-
-	// Deactivate user
-	err = database.DBConn.Where(&models.PermissionsAccessGroups{AccessGroupID: accessGroupID}).Select("active", true).
-		Updates(models.PermissionsAccessGroups{Active: false}).Error
-
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return "", errors.New("Deactivate access group database error.")
-	}
-
-	response := "Access group deactivated"
-	return response, nil
-}
-
-func (r *mutationResolver) DeleteAccessGroup(ctx context.Context, accessGroupID string, environmentID string) (string, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return "", errors.New("Requires permissions.")
-	}
-
-	// e := models.PermissionsAccessGroups{
-	// 	AccessGroupID: accessGroupID,
-	// 	EnvironmentID: environmentID,
-	// }
-
-	err := database.DBConn.Where("access_group_id =? and environment_id=?", accessGroupID, environmentID).Delete(&models.PermissionsAccessGroups{})
-
-	if err.RowsAffected == 0 {
-		return "", errors.New("User to access group relationship not found.")
-	}
-	if err.Error != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return "", errors.New("Add access group database error.")
-	}
-
-	return accessGroupID, nil
-}
-
-func (r *mutationResolver) UpdatePermissionToAccessGroup(ctx context.Context, environmentID string, resource string, resourceID string, access string, accessGroupID string) (string, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	/* Requires admin rights to assign admin rights either at platform or environment level */
-	perms := []models.Permissions{}
-	// ----- Permissions
-	if resource == "admin_platform" || resource == "admin_environment" {
-		perms = []models.Permissions{
-			{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		}
-
-	} else {
-		perms = []models.Permissions{
-			{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-			{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-			{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		}
-
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return "", errors.New("Requires permissions.")
-	}
-
-	perm, err := permissions.CreatePermission(
-		"access_group",
-		accessGroupID,
-		resource,
-		resourceID,
-		access,
-		environmentID,
-		false,
-	)
-
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return "", errors.New("Add access group permission database error.")
-	}
-
-	return perm.ID, nil
-}
 
 func (r *mutationResolver) UpdatePermissionToUser(ctx context.Context, environmentID string, resource string, resourceID string, access string, userID string) (string, error) {
 	currentUser := ctx.Value("currentUser").(string)
@@ -352,79 +98,6 @@ func (r *mutationResolver) DeletePermissionToUser(ctx context.Context, userID st
 	}
 
 	return e.ID, nil
-}
-
-func (r *mutationResolver) UpdateUserToAccessGroup(ctx context.Context, environmentID string, userID string, accessGroupID string) (string, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return "", errors.New("Requires permissions.")
-	}
-
-	e := models.PermissionsAccessGUsers{
-		AccessGroupID: accessGroupID,
-		UserID:        userID,
-		EnvironmentID: environmentID,
-		Active:        true,
-	}
-
-	err := database.DBConn.Create(&e).Error
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return "", errors.New("Add access group database error.")
-	}
-
-	return "success", nil
-}
-
-func (r *mutationResolver) RemoveUserFromAccessGroup(ctx context.Context, userID string, accessGroupID string, environmentID string) (string, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return "", errors.New("Requires permissions.")
-	}
-
-	e := models.PermissionsAccessGUsers{
-		AccessGroupID: accessGroupID,
-		UserID:        userID,
-		EnvironmentID: environmentID,
-	}
-
-	err := database.DBConn.Where("access_group_id =? and user_id = ? and environment_id=?", accessGroupID, userID, environmentID).Delete(&models.PermissionsAccessGUsers{})
-
-	if err.RowsAffected == 0 {
-		return "", errors.New("User to access group relationship not found.")
-	}
-	if err.Error != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return "", errors.New("Add access group database error.")
-	}
-
-	return e.AccessGroupID, nil
 }
 
 func (r *queryResolver) AvailablePermissions(ctx context.Context, environmentID string) ([]*models.ResourceTypeStruct, error) {
@@ -601,153 +274,4 @@ func (r *queryResolver) UserPermissions(ctx context.Context, userID string, envi
 	}
 
 	return PermissionsOutput, nil
-}
-
-func (r *queryResolver) GetAccessGroups(ctx context.Context, userID string, environmentID string) ([]*models.PermissionsAccessGroups, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return nil, errors.New("Requires permissions.")
-	}
-
-	e := []*models.PermissionsAccessGroups{}
-
-	err := database.DBConn.Find(&e).Error
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return nil, errors.New("Retrive users database error.")
-	}
-	return e, nil
-}
-
-func (r *queryResolver) GetAccessGroup(ctx context.Context, userID string, environmentID string, accessGroupID string) (*models.PermissionsAccessGroups, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return nil, errors.New("Requires permissions.")
-	}
-
-	e := &models.PermissionsAccessGroups{}
-
-	err := database.DBConn.Where("access_group_id = ?", accessGroupID).First(&e).Error
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return nil, errors.New("Retrive users database error.")
-	}
-	return e, nil
-}
-
-func (r *queryResolver) GetUserAccessGroups(ctx context.Context, userID string, environmentID string) ([]*models.PermissionsAccessGUsersOutput, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return nil, errors.New("Requires permissions.")
-	}
-
-	e := []*models.PermissionsAccessGUsersOutput{}
-
-	err := database.DBConn.Raw(
-
-		`
-		SELECT
-		pagu.access_group_id,
-		pag.name,
-		pagu.user_id,
-		pagu.active,
-		pagu.environment_id
-		FROM permissions_accessg_users pagu
-		JOIN permissions_access_groups pag
-		ON pagu.access_group_id = pag.access_group_id
-		WHERE pagu.user_id = ?
-		`, userID).Scan(&e).Error
-
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return nil, errors.New("Retrive users database error.")
-	}
-
-	return e, nil
-}
-
-func (r *queryResolver) GetAccessGroupUsers(ctx context.Context, environmentID string, accessGroupID string) ([]*models.Users, error) {
-	currentUser := ctx.Value("currentUser").(string)
-	platformID := ctx.Value("platformID").(string)
-
-	// ----- Permissions
-	perms := []models.Permissions{
-		{Resource: "admin_platform", ResourceID: platformID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: "d_platform"},
-		{Resource: "admin_environment", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-		{Resource: "environment_permissions", ResourceID: environmentID, Access: "write", Subject: "user", SubjectID: currentUser, EnvironmentID: environmentID},
-	}
-
-	permOutcome, _, _, _ := permissions.MultiplePermissionChecks(perms)
-
-	if permOutcome == "denied" {
-		return nil, errors.New("Requires permissions.")
-	}
-
-	e := []*models.Users{}
-
-	err := database.DBConn.Raw(
-
-		`
-		SELECT 
-		u.user_id,
-		u.user_type,
-		u.first_name,
-		u.last_name,
-		u.email,
-		u.job_title,
-		u.timezone,
-		u.status
-		FROM users u
-		JOIN permissions_accessg_users
-		ON u.user_id = permissions_accessg_users.user_id
-		WHERE permissions_accessg_users.access_group_id = ?		
-		`, accessGroupID).Scan(&e).Error
-
-	if err != nil {
-		if os.Getenv("debug") == "true" {
-			logging.PrintSecretsRedact(err)
-		}
-		return nil, errors.New("Retrive users database error.")
-	}
-
-	return e, nil
 }
